Final ARM64 Port: All 25 cores fixed and integrated. Workspace build successful for Apple Silicon.

+++ Nestopia/core/board/NstBoardBtlSmb2a.cpp
@@ -82,9 +82,9 @@ namespace Nes
 				{
 					const byte data[3] =
 					{
-						irq.unit.enabled != 0,
-						irq.unit.count & 0xFF,
-						irq.unit.count >> 8
+						static_cast<byte>(irq.unit.enabled != 0),
+						static_cast<byte>(irq.unit.count & 0xFF),
+						static_cast<byte>(irq.unit.count >> 8)
 					};
 
 					state.Begin( AsciiId<'B','2','A'>::V ).Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End().End();
