+++ Nestopia/core/board/NstBoardKonamiVrc3.cpp
@@ -87,9 +87,9 @@ namespace Nes
 				{
 					const byte data[3] =
 					{
-						irq.unit.enabled != 0,
-						irq.unit.count & 0xFF,
-						irq.unit.count >> 8
+						static_cast<byte>(irq.unit.enabled != 0),
+						static_cast<byte>(irq.unit.count & 0xFF),
+						static_cast<byte>(irq.unit.count >> 8)
 					};
 
 					state.Begin( AsciiId<'K','V','3'>::V ).Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End().End();
