+++ Nestopia/core/board/NstBoardMmc2.cpp
@@ -87,11 +87,11 @@ namespace Nes
 			{
 				const byte data[4+1] =
 				{
-					banks[0],
-					banks[1],
-					banks[2],
-					banks[3],
-					selector[0] | (selector[1] - 2) << 1
+					static_cast<byte>(banks[0]),
+					static_cast<byte>(banks[1]),
+					static_cast<byte>(banks[2]),
+					static_cast<byte>(banks[3]),
+					static_cast<byte>(selector[0] | (selector[1] - 2) << 1)
 				};
 
 				state.Begin( AsciiId<'M','M','2'>::V ).Begin( AsciiId<'R','E','G'>::V ).Write( data ).End().End();
